Fixed a variety of bugs that may or may not actually have an impact on the final library file.  Not sure, but the changes were made for good style and ensure a consistent product.
In addition, some functions in TimeUtly were simplified,
but they should have the same results.


git-svn-id: svn+ssh://localhost/home/bvr/svnrepos/Programs/BUtils@10 c770f1bc-0775-48ed-bd74-6edc8452e29f

+++ trunk/src/CmdLineUtly.C
@@ -27,7 +27,7 @@ vector <string> ProcessFlatCommandLine(int ArgumentCount, char *ArgLine[])
 	vector <string> CmdLineStrings(ArgumentCount - 1, "");	
 	// string vector of size one less than the size of ArgLine filled with nothing
 
-	for (int ArgIndex = 0; ArgIndex < CmdLineStrings.size(); ArgIndex++)
+	for (size_t ArgIndex = 0; ArgIndex < CmdLineStrings.size(); ArgIndex++)
 	{
 		CmdLineStrings[ArgIndex] = ArgLine[ArgIndex + 1];
 	}
